Added torkill for Linux

+++ OnionSwitch_GUI/OnionSwitch_TorCheck.py
@@ -20,6 +20,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import threading
 import os
+import signal
+from subprocess import check_output
 
 
 import stem.process
@@ -64,7 +66,7 @@ class TorCheck(object):
         except Exception:
             TorCheck.ended_successfull = False
 
-    def CheckTor(self, countrycode):
+    def CheckTor(self, countrycode, platform):
 
         try:
             # If tor.exe is already started, quit it and start a new thread to
@@ -75,13 +77,19 @@ class TorCheck(object):
 
             if countrycode != "":
                 TorCheck.connected = False
-                tasks = os.popen('tasklist').readlines()
 
-                for task in tasks:
-                    tor_index = task.find("tor.exe")
-                    if tor_index != -1:
-                        if tor_index == 0:
-                            os.system("taskkill /f /im tor.exe")
+                if platform == "Windows":
+                    tasks = os.popen('tasklist').readlines()
+
+                    for task in tasks:
+                        tor_index = task.find("tor.exe")
+                        if tor_index != -1:
+                            if tor_index == 0:
+                                os.system("taskkill /f /im tor.exe")
+
+                if platform == "Linux":
+                    pid = check_output(["pidof", "tor"])
+                    os.kill(int(pid), signal.SIGKILL)
 
                 torcheck_thread = threading.Thread(
                     target=TorCheck.CheckNode, args=(
@@ -92,11 +100,17 @@ class TorCheck(object):
                 torcheck_thread.join(timeout=10)
 
                 if TorCheck.ended_successfull is False:
-                    tasks = os.popen('tasklist').readlines()
+                    if platform == "Windows":
+                        tasks = os.popen('tasklist').readlines()
 
-                    for task in tasks:
-                        tor_index = task.find("tor.exe")
-                        if tor_index != -1:
-                            os.system("taskkill /f /im tor.exe")
+                        for task in tasks:
+                            tor_index = task.find("tor.exe")
+                            if tor_index != -1:
+                                if tor_index == 0:
+                                    os.system("taskkill /f /im tor.exe")
+
+                    if platform == "Linux":
+                        pid = check_output(["pidof", "tor"])
+                        os.kill(int(pid), signal.SIGKILL)
         except Exception as exc:
             print(exc)
